+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/read-case-flag-field.component.ts
@@ -6,7 +6,7 @@ import { CaseFlagStateService } from '../../case-editor/services/case-flag-state
 import { AbstractFieldReadComponent } from '../base-field/abstract-field-read.component';
 import { PaletteContext } from '../base-field/palette-context.enum';
 import { FlagDetailDisplay, FlagsWithFormGroupPath } from './domain';
-import { CaseFlagStatus, CaseFlagSummaryListDisplayMode } from './enums';
+import { CaseFlagDisplayContextParameter, CaseFlagStatus } from './enums';
 
 @Component({
   selector: 'ccd-read-case-flag-field',
@@ -19,17 +19,12 @@ export class ReadCaseFlagFieldComponent extends AbstractFieldReadComponent imple
   public caseLevelCaseFlagData: FlagsWithFormGroupPath;
   public paletteContext = PaletteContext;
   public flagForSummaryDisplay: FlagDetailDisplay;
-  public summaryListDisplayMode: CaseFlagSummaryListDisplayMode;
   public displayContextParameter: string;
   public caseFlagsExternalUser = false;
   public pathToFlagsFormGroup: string;
   public readonly caseLevelCaseFlagsFieldId = 'caseFlags';
   public readonly caseNameMissing = 'Case name missing';
   public readonly readSupportMode = '#ARGUMENT(READ,EXTERNAL)';
-  private readonly createMode = '#ARGUMENT(CREATE)';
-  private readonly createExternalMode = '#ARGUMENT(CREATE,EXTERNAL)';
-  private readonly updateMode = '#ARGUMENT(UPDATE)';
-  private readonly updateExternalMode = '#ARGUMENT(UPDATE,EXTERNAL)';
 
   constructor(
     private readonly route: ActivatedRoute,
@@ -76,18 +71,16 @@ export class ReadCaseFlagFieldComponent extends AbstractFieldReadComponent imple
       if (flagLauncherComponent) {
         // The FlagLauncher component holds a reference (selectedFlagsLocation) containing the CaseField instance to
         // which the new flag has been added
-        if ((flagLauncherComponent.caseField.display_context_parameter === this.createMode ||
-            flagLauncherComponent.caseField.display_context_parameter === this.createExternalMode)
+        if ((flagLauncherComponent.caseField.display_context_parameter === CaseFlagDisplayContextParameter.CREATE ||
+            flagLauncherComponent.caseField.display_context_parameter === CaseFlagDisplayContextParameter.CREATE_EXTERNAL)
             && flagLauncherComponent.selectedFlagsLocation) {
           this.pathToFlagsFormGroup = flagLauncherComponent.selectedFlagsLocation.pathToFlagsFormGroup;
           this.flagForSummaryDisplay = this.extractNewFlagToFlagDetailDisplayObject(
             flagLauncherComponent.selectedFlagsLocation);
-          // Set the display mode for the "Review flag details" summary page
-          this.summaryListDisplayMode = CaseFlagSummaryListDisplayMode.CREATE;
         // The FlagLauncher component holds a reference (selectedFlag), which gets set after the selection step of the
         // Manage Case Flags journey
-        } else if ((flagLauncherComponent.caseField.display_context_parameter === this.updateMode ||
-          flagLauncherComponent.caseField.display_context_parameter === this.updateExternalMode) &&
+        } else if ((flagLauncherComponent.caseField.display_context_parameter === CaseFlagDisplayContextParameter.UPDATE ||
+          flagLauncherComponent.caseField.display_context_parameter === CaseFlagDisplayContextParameter.UPDATE_EXTERNAL) &&
           flagLauncherComponent.selectedFlag) {
             this.flagForSummaryDisplay =
               this.formGroup.get(flagLauncherControlName)['component'].selectedFlag.flagDetailDisplay;
@@ -101,8 +94,6 @@ export class ReadCaseFlagFieldComponent extends AbstractFieldReadComponent imple
               ...caseFlagFormGroupValue
             };
           }
-          // Set the display mode for the "Review flag details" summary page
-          this.summaryListDisplayMode = CaseFlagSummaryListDisplayMode.MANAGE;
         }
       }
     }