+++ RELEASE-NOTES.md
@@ -4,6 +4,9 @@
 **EUI-8070** Display "decision reason" (flag update comment) for "Not approved" flags to internal HMCTS users only
 **EUI-8071** Hide "Active flags" banner from external users
 
+### Version 6.10.7-case-flags-legal-rep-comments-page-v3
+**EUI-7373** Reasonable Adjustments legal rep display and edit ‘Tell us why the support is no longer needed' screen
+
 ### Version 6.10.7-case-flags-welsh-language-support
 **EUI-7858** Ensure support for Welsh language toggle by storing flag description and comments entries in correct `_cy` fields, and displaying flag comments according to language selection